Make hour/minute/second/timestamp work in sequence of text
