+++ languages/objecten/languageModels/gegevensspraak/behavior.mps
@@ -20538,6 +20538,7 @@
         <ref role="2sp9C9" to="tpee:fKQsSyN" resolve="EnumConstantDeclaration" />
       </node>
     </node>
+    <node concept="2tJIrI" id="3RxOc3uTld4" role="jymVt" />
     <node concept="2YIFZL" id="62XJbzXeKIV" role="jymVt">
       <property role="TrG5h" value="XmlRegexVoorGranulariteit" />
       <node concept="3clFbS" id="62XJbzXeKIX" role="3clF47">
